Authentication

+++ hotels-backend/Hotels/Controllers/ReservationsController.cs
@@ -6,6 +6,9 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Hotels.Models;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
+using Hotels.Extensions;
 
 namespace Hotels.Controllers
 {
@@ -22,6 +25,7 @@ namespace Hotels.Controllers
         }
 
         // GET: api/Reservations
+        [Authorize (Roles = "Admin")]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Reservation>>> GetReservations([FromQuery] int page)
         {
@@ -32,6 +36,7 @@ namespace Hotels.Controllers
         }
 
         // GET: api/Reservations/5
+        [Authorize]
         [HttpGet("{id}")]
         public async Task<ActionResult<Reservation>> GetReservation(int id)
         {
@@ -42,63 +47,96 @@ namespace Hotels.Controllers
                 return NotFound();
             }
 
-            return reservation;
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+
+            if (identity.GetAuthorizedUserId() == reservation.UserId || identity.GetAuthorizedUserRole() == "Admin")
+            {                
+                return reservation;
+            }
+
+            return Forbid();
         }
 
         // PUT: api/Reservations/5
+        [Authorize]
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutReservation(int id, Reservation reservation)
+        public async Task<IActionResult> PutReservation(int id, Reservation modifiedReservation)
         {
-            if (id != reservation.Id)
+            if (id != modifiedReservation.Id)
             {
                 return BadRequest();
             }
 
-            _context.Entry(reservation).State = EntityState.Modified;
+            var reservation = await _context.Reservations.FindAsync(id);
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
 
-            try
+            if (identity.GetAuthorizedUserId() == reservation.UserId || identity.GetAuthorizedUserRole() == "Admin")
             {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!ReservationExists(id))
+                _context.Entry(modifiedReservation).State = EntityState.Modified;
+
+                try
                 {
-                    return NotFound();
+                    await _context.SaveChangesAsync();
                 }
-                else
+                catch (DbUpdateConcurrencyException)
                 {
-                    return Conflict();
+                    if (!ReservationExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        return Conflict();
+                    }
                 }
+
+                return NoContent();
             }
 
-            return NoContent();
+            return Forbid();
         }
 
         // POST: api/Reservations
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<Reservation>> PostReservation(Reservation reservation)
         {
-            _context.Reservations.Add(reservation);
-            await _context.SaveChangesAsync();
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
 
-            return CreatedAtAction("GetReservation", new { id = reservation.Id }, reservation);
+            if (identity.GetAuthorizedUserId() == reservation.UserId || identity.GetAuthorizedUserRole() == "Admin")
+            {
+                _context.Reservations.Add(reservation);
+                await _context.SaveChangesAsync();
+
+                return CreatedAtAction("GetReservation", new { id = reservation.Id }, reservation);
+            }
+
+            return Forbid();
         }
 
         // DELETE: api/Reservations/5
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteReservation(int id)
         {
             var reservation = await _context.Reservations.FindAsync(id);
+
             if (reservation == null)
             {
                 return NotFound();
             }
 
-            _context.Reservations.Remove(reservation);
-            await _context.SaveChangesAsync();
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
 
-            return NoContent();
+            if (identity.GetAuthorizedUserId() == reservation.UserId || identity.GetAuthorizedUserRole() == "Admin")
+            {
+                _context.Reservations.Remove(reservation);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
+            }
+
+            return Forbid();          
         }
 
         private bool ReservationExists(int id)
